+++ plugins/source/azure/resources/services/cognitiveservices/accounts_mock_test.go
@@ -34,7 +34,27 @@ func createAccounts(router *mux.Router) error {
 		}
 	})
 
-	return nil
+	if err := createAccountDeployments(router); err != nil {
+		return err
+	}
+
+	if err := createAccountPrivateEndpointConnections(router); err != nil {
+		return err
+	}
+
+	if err := createAccountPrivateLinkResources(router); err != nil {
+		return err
+	}
+
+	if err := createAccountModels(router); err != nil {
+		return err
+	}
+
+	if err := createAccountUsages(router); err != nil {
+		return err
+	}
+
+	return createAccountSKUs(router)
 }
 
 func TestAccounts(t *testing.T) {